functions/pyproject.toml
[tool.poetry]
name = "ticket-guard-functions"
version = "0.1.0"
description = "Azure Functions for Ticket Guard"
authors = ["Shuji Miyoshi <shuji.miyoshi@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
azure-functions = "1.18.0"

[tool.poetry.dev-dependencies]
mypy = "1.11.2"
pytest = "8.3.3"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
